+++ src/index.ts
@@ -34,6 +34,7 @@ export * from "./components/Sparkline";
 export * from "./components/HeartbeatSparkline";
 export * from "./components/TrendSparkline";
 export * from "./components/DistributionSparkline";
+export * from "./components/RateGauge";
 export * from "./components/LiveHeartbeatTrace";
 export * from "./components/ConnectionStatus";
 export * from "./components/ConversationTree";